Add modular callback classes for TabNet training
Refactored callbacks into separate modular files: `callback.py`, `container.py`, `early_stopping.py`, `history.py`, and `lr_scheduler.py`. Improved code organization and maintainability by isolating functionality for individual callbacks. Simplified imports in `__init__.py` for easier access.
